requirements.txt
python-nmap==0.6.1
IPy==1.0